requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
chardet==4.0.0
dataclasses==0.6
deprecation==2.1.0
Django==4.2.6
django-ckeditor==6.7.0
django-js-asset==2.1.0
gotrue==0.2.0
h11==0.14.0
httpcore==0.12.3
httpx==0.16.1
idna==2.10
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
postgrest-py==0.4.0
psycopg2-binary==2.9.9
py==1.11.0
pytest==6.2.5
python-dateutil==2.8.2
realtime-py==0.1.3
requests==2.25.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
supabase-py==0.0.2
toml==0.10.2
typing_extensions==4.8.0
urllib3==1.26.17
uuid==1.30
websockets==9.1
